Update doc on importing data
I've converted the numbered list of steps into titles so that we
can have a table of contents and ability to link to each step.
Hopefully it's makes it easier to follow and less intimidating
for those less familiar to mapit.

I've also updated the steps with more information, and removed
some things that are no longer relevant.