Update documentation
Some refining of the documentation following the recent data import.

I have also introduced a new file, HISTORY.md, where we can log
some of the main changes for each update. This wil be useful when
we try to import data in future and are aware of anything that might
affect the next import.