Only run Github Actions for Python 3.6
Mysociety runs GA for several Python versions, but this isn't
necessary for us as we are only using 3.6, and test with Django 3.1
which is the latest version supported by Mapit.

We are also using our `requirements.txt` file so that we will always
be testing against the same dependency versions we as we use in
production.

Add Raven