Disable test
This test is failing in CI, and the method within the test is only
used when importing data from source. As we only import data locally
we don't have a need for the test.

I've disabled it rather than removing it completely for reference,
and we may get conflicts in future when we pull upstream changes
from Mysociety again.